+++ common/src/main/java/com/feed_the_beast/ftbquests/quest/reward/Reward.java
@@ -74,7 +74,7 @@ public abstract class Reward extends QuestObjectBase
 	}
 
 	@Override
-	public final int getParentID()
+	public final long getParentID()
 	{
 		return quest.id;
 	}
